move isNative to where I meant to put it
