+++ src/apps/halofi/positions.ts
@@ -9,7 +9,7 @@ import { getCompatibleGamesFromAPI } from './haloFiApi'
 import { PlayerStructIndex, getPlayerStructFromGames } from './haloFiContract'
 import { NetworkId } from '../../api/networkId'
 import { getTokenId } from '../../runtime/getTokenId'
-import { isNative } from '../runtime/isNative'
+import { isNative } from '../../runtime/isNative'
 
 const hook: PositionsHook = {
   getInfo() {
