+++ src/runtime/isNative.ts
@@ -1,4 +1,4 @@
-import { NetworkId } from '../../api/networkId'
+import { NetworkId } from '../api/networkId'
 
 const ETHER_HEX_IDENTIFIER = '0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee' // curve and some other dapps use this address as an identifier for the native asset
